move fim to it's own file
